zebra interfaces NB conversion

+++ lib/if.h
@@ -195,7 +195,7 @@ struct if_link_params {
 	uint32_t min_delay;		/* Link Min Delay */
 	uint32_t max_delay;		/* Link Max Delay */
 	uint32_t delay_var;		/* Link Delay Variation */
-	float pkt_loss;			/* Link Packet Loss */
+	uint32_t pkt_loss;		/* Link Packet Loss */
 	float res_bw;			/* Residual Bandwidth */
 	float ava_bw;			/* Available Bandwidth */
 	float use_bw;			/* Utilized Bandwidth */
